+++ lib/wootricsdk_flutter.dart
@@ -147,4 +147,21 @@ class WootricsdkFlutter {
   static showSurveyWithEvent(String eventName) {
     WootricsdkFlutterPlatform.instance.showSurveyWithEvent(eventName);
   }
+
+  ///CUSTOM
+  /// For promoters (score 9-10)
+  /// Display a custom thank you message (setPromoterThankYouMessage)
+  /// Display a custom thank you button and link redirect (setPromoterThankYouLinkWithText)
+  static setPromoterCustomThankYouScreen(
+      String message, String label, String url) {
+    WootricsdkFlutterPlatform.instance
+        .setPromoterCustomThankYouScreen(message, label, url);
+  }
+
+  ///CUSTOM
+  ///Override survey theme with custom colors for primary and secondary colors.
+  ///To set a custom color pass appropriate [primary] and [secondary] in hex format.
+  static setSurveyTheme(String primary, {String? secondary}) {
+    WootricsdkFlutterPlatform.instance.setSurveyColors(primary, secondary);
+  }
 }